requirements.txt
bcrypt
celery
email-validator
Flask
Flask-CORS
flask-restx
Flask-SQLAlchemy
gunicorn
jsonschema
marshmallow
psutil
PyJWT
PyMySQL
python-dateutil
python-dotenv
redis
requests
Werkzeug